Remove per-transport settings
Remove the various `${transport}_delivery_slot_cost`, etc. settings as
they are dynamically defined based on whatever servers are defined. Keep
the various `default_*` settings but require the use of the new
`postfix::main` defined type to set the per-transport equivalents.

Add autorequire logic to the `postfix_main` type for any of these
per-transport settings to require the service in question.